Module connections working
