--- src/lib.rs
@@ -1,12 +1,0 @@
-mod controllers;
-
-use controllers::create_routes;
-
-pub async fn run() {
-    let app = create_routes();
-
-    axum::Server::bind(&"127.0.0.1:3000".parse().unwrap())
-        .serve(app.into_make_service())
-        .await
-        .unwrap();
-}
