+++ src/views.rs
@@ -1,1 +1,2 @@
-mod views;
+pub mod sys_info_view;
+
