+++ src/views/sys_info_view.rs
@@ -1,0 +1,7 @@
+// template logic for sys_view
+//use axum::Json;
+//use serde::Serialize;
+
+pub fn sys_info_template() {
+    println!("GET sysinfo view TEST Hello World !!post from my own file");
+}
